crawler_bqjr/requirements.txt
six
cchardet
ujson
urllib3
requests
lxml
scrapy
selenium
beautifulsoup4
pymongo
pymysql
pika
python-dateutil
pytz
xlrd
imapclient
django
django-oauth-toolkit
django-cors-middleware
rsa
numpy
opencv-contrib-python
Pillow
piltesseract
pytesseract
baidu-aip
pycryptodome
psutil
filelock
Cython
spp
pywin32